Modified app structure and database
>changed app structure
changed the structure by pushing all the code into new 'flaskstruct'
direcctory to import another py files. Changed app.py to __init__.py
and imported this to external app.py file as
>from flaskstuct import app

>database addition(flask-mysqld)
made another database config file dbconnect.py put config code to it
and a connection function which can be called when we need db
connection. Imported this dbconnect.py to __init__.py as
>from flaskstruct.dbconnect import connection

>moved routes to routes.py
Imported this to __init__.py as
>from flaskstruct import routes